+++ src/components/common/Card/Card.tsx
@@ -1,4 +1,4 @@
-import { Card as MuiCard, CardProps as MuiCardProps } from "@mui/material";
+import MuiCard, { type CardProps as MuiCardProps } from "@mui/material/Card";
 
 export interface CardProps extends MuiCardProps {
   /**